Add an env variable to avoid locking the credentials directory
This commit allows disabling locking the credentials directory when
V23_CREDENTIALS_NO_LOCK environment variable is defined. This is useful when
the credentials are stored in a read-only location.

A realistic use case is when a Vanadium program is package in a Docker image
that doesn't have v23agentd and the credentials are mapped read-only using the
`docker run -v ...:ro`. In this case both V23_CREDENTIALS_NO_AGENT and
V23_CREDENTIALS_NO_LOCK need to be set. This can be done either in the
Dockerfile or via:

  docker run \
    --env V23_CREDENTIALS_NO_AGENT=1 \
    --env V23_CREDENTIALS_NO_LOCK=1 \
    ...