v.io/x/ref/runtime: fix a bug affecting restarting of servers on the same host+port
Every Vanadium processes uses a different RID (Routing ID) to ensure
that clients can distinguish server instances. If a server uses
a fixed address and port, then clients of that server will
see different RIDs after the server restarts. The current code
assumes that this case of a changed RID implies the use of a proxy.
This revision changes this assumption so that a different RID on the
same address and port is not assumed to be a proxy. The old
behaviour leads to unexpected behaviour, and often to authentication
failures due to the server being treated as a proxy, in particular,
the client fails to send any matching blessings to the server (since
proxies should not be sent blessings intended for the servers
they are proxying).