Expose the real external address of a peer for a Vanadium call
The v23/rpc/server/ServerCall interface doesn't expose via the
RemoteEndpoint() the real external address of the peer. This change
adds a RemoteAddr() method that does that.

This commit also turns off the fancy output from gradlew to make it
easier to see the failures for the Java tests (which seem to be flaky on
Travis CI).