Fix a fragmentation bug exhibit by xproxyd
The existing logic to avoid fragmentation from flw.writeMsg doesn't
always work causing the xproxy to send to peers (either clients or
servers) incorrectly fragmented messages. As far as I can tell this
happens because the xproxy uses asymmetric connections: the one from
clients/servers to xproxy is *conn.flw while the one from xproxy to
clients/server is tcputil.tcpConn. This, in turn, is causing the
Conn.setup from x/ref/runtime/internal/flow/conn/auth.go to not be able
to detect that encryption is disabled which is the trigger used for
allowing or not the fragmention.

The solution in this commit is to add a dedicated function to disable
fragmentation and use it in the xproxyd code. I did not use chose to
export and use the existing flw.disableEncryption because in this case
we are not actually interested to disable the encryption.