v23/context: add FromGoContext to allow for conversion to/from context.Context" (#48)
This allows converting a context.Context of unknown origin to vanadium
*Context.T.

Background: if a library (outside v.io) is given a *vcontext.T, and it wraps the
context around context.{WithCancel,WithDeadline,WithValue}, then the context can
no longer be passed back to vanadium functions. This change allows for such
conversions.