x/ref/runtime/internal: fix flaky tests due to rare race conditions in the networking layer (#55)
There are three race conditions that are addressed by this pull request that would show themselves in rare failures for the TestAuthorizationDuringResolve (naming/namespace) test when the tests are run on a heavily loaded. It was possible to repro them using as well as seeing them on travis at peak load times during the day, but not evenings/weekends.

$ seq 1 2000 | parallel -j 30 'go test -v -run=TestAuthorizationDuringResolve -count=1

dialing a side channel via DialSideChannel is racy in that it relies on finding the same connection, in the cache, as the existing parent connection (see rpc/client.go:479 and below). The connection cache may end up with multiple entries for the same remote endpoint depending on relative arrival of messages, but the existing code would return only the first authorized one connection. The change made is to change the API to supply the existing connection (DialSideChannelCached) and to search the connection cache for that connection as expected by the cal site (rpc/client.go:479). The paranoid, though strictly unnecessary, checks at the call site are kept for now.
With the above problem fixed, two further ones showed themselves. The first was a deadlock that would result in the test itself timing out. The second, also a deadlock, that would result in a name resolution timeout.

the first deadlock occurs because the race causes a connection close to hang which prevents the test from cleaning up after itself. The race is in NewAccepted, which uses a background goroutine to call acceptHandshake but waits for that goroutine to return, however, it is possible for the waiting select to timeout, or for its context to be cancelled, in which case the connection is closed via its Close method see, conn.go:358. However calling Close directly will cause it block on the loopWG wait group, which is only released when the prior background goroutine finishes, which may take a long time or forever if it's blocked on reading a connection that is never closed. The fix is to extend internalClose to not block on the loopWG waitgroup in this case.
NewAccepted calls acceptHandshake, but if that that times out or the call is cancelled, then it doesn't make sense to wait for it to finish, hence the change to conn.go:363, to close the connection but without on the loopWG waitgroup.

the second timeout occurs due to a bug in the handling of OpenFlow messages via the NewAccepted method again. NewAccepted calls acceptHandshake and then remoteAuth to complete the authentication handshake. However remoteAuth calls handleMessage for all message types, including OpenFlow messages. handleMessage will call the flow handler supplied to it to pass the openflow message back to the flow manager. This has the potential to block, and at least for the call to NewAccepted from the flow manager that blocked call can only be unblocked when NewAccepted returns - hence the deadlock. The fix is to call handleMessage asynchronously in this case as is the normal code path (via readLoop). See manager.go:565 for the call site, not that the flow handler has a channel that HandleFlow will block on, but that that channel is not closed until NewAccepted returns (line 625).