x/ref/runtime: ensure that command line flags are parsed (#63)
PR #52 refactored factory initialization but broken flag parsing for applications that didn't use the cmdline package. This PR fixes that and also makes it possible for applications to register the factory flags with on their flag set, it turns out that this is used by some applications at GRAIL (eg. reflow).