x/ref/runtime/factories: changes to support cross-compilation (#69)
The roaming profile uses x/lib/osnetconfig to detect and read networking changes. osnetconfig requires the use of cgo which implies that any user of the roaming profile cannot be cross-compiled. This PR uses build constraints to disable the use of osnetconfig when cross compiling.
In addition, a new factory, static, is provided that explicitly does not enable roaming and it used where appropriate.