add go.mod and go.sum in preparation for tagging the first v0.1 release (#71)
Add go.mod and go.sum in preparation for tagging the first v0.1 release of v.io/v23 and v.io/x/ref as a go module. This repository will be tagged as v0.1.0 after this lands.
Note, that it any test/app that issues a 'go run' or 'go get' command from outside of a go tree with go.mod needs to disable module support by setting GO111MODULE to no.