v23/vom: add the ability to re-encode an already encoded streams (#80)
This PR adds the ability to splice an already encoded vom stream into another. This is required because of how the RPC system is currently structured in that it creates a new vom stream for each call internally to encode the service and method names etc. With this change it is possible to create a vom stream using an alternate implementation (say python) and merge it into the RPC created one in a vdl generated or handwritten stub. For example:

merge := vom.MergeEncodedBytes{Data: rpcargs}
cltCall, err := clt.StartCall(ctx, name, method, []interface{}{merge})
....
extractor := vom.ExtractEncodedBytes{}
err := cltCall.Finish(&extractor)