lint: fix lint 'unused' and 'gosimple' errors (#114)
This PR fixes 'unused' and 'gosimple' errors. Much of the code is kept but commented with nolint since it's either work in progress or unclear if it should be kept. code that is obviously dead is removed. Some things to note:

- the vdl code generation is changed to emit gosimple compatible code 
  (x/ref/lib/vdl/codegen/golang).
- the v.io/x/lib dependency is upgraded to 0.1.5 since that version is the first that passes lint
  tests and has trimmed dependencies.

In addition two bugs were caught and fixed:

- v23/query/engine/internal/query_functions/query_functions.go: the code looks as if was
expecting break to exit out of the outer for loop and not the inner switch statement.
- vom/encoder.go:459, missing type in case arm.