v23/context: use context.Context to implement this package (#145)
This PR reworks the implementation of v23/context to directly use context.Context to implement its functionality. This both simplifies the implementation of this package and makes the differences to context.Context easier to understand. The essential difference being support for asynchronous processing whose cancelation is detached from its immediate parent but not from a 'root' context is intended to be canceled on process termination. This is similar to the concept of context.Background() except that all state (ie. values present in the immediate parent context and its ancestors) are copied to the new context. This is convenient for vanadium where security credentials etc are carried across API boundaries as values in the context.