x/ref/lib/security: use read-only lock for reading/loading principal data. (#150)
This PR changes the locking strategy to use a read-only lock for read/loading principals versus a read-write lock for mutation. This should allow for servers to be run with read-only filesystems hosting their credentials. It also adds an environment variable V23_CREDENTIALS_READONLY_FILESYSTEM to specify when credentials are on a read-only filesystem, the previous V23_CREDENTIALS_NO_LOCK is now an alias for the former and will be deleted in the future.