v23/verror: make verror easier to use and compatible with the new errors.Is/Unwrap APIs (#149)
See Issue #151 for a more complete description of the issues this PR is addressing. This PR implements the revised verror API, updates all code in vanadium/core to use that API, but does not remove the old API.

verror.{NewID,NewIDAction} are added to allow the convenient creation of IDActions. IDActions now have methods for generating errors, i.e. Errorf which mimics fmt.Errorf, and Message which accepts a pre-formatted message.
errors.Unwrap and errors.Is are now supported, IDAction and SubErr now implement error to support their use
verror.WithSubErrors is added to generalize the association of subErrors with an existing verror.E
all existing code is updated to use these API extensions.
Subsequent PRs will remove the i18n support and older API.