x/ref/lib/vdl: update vdl error handling (#156)
This PR continues to address #149 and updates the vdl language spec as follows:

the error definition no longer needs an i18n format, warnings are issued if one is given, and also a warning is given if one is not to use the new APIs
- new functions Errorf... and Messagef are generated to create errors
- new functions Params... are generated to safely extract the defined parameters from an error
- all code in core is updated to use the new APIs, but the old APIs are still generated

The vdl command offers two new options to aid with the transition to the new API. The first, -show-warnings, can be used to disable printing of the warnings mentioned above. Support for the old API can be removed with, --errors-no-i18n=true, which for now defaults to false.