x/ref/internal: fix a circular module dependency between internal/logger and v.io/v23 (#161)
PR #148 created a new submodule (x/ref/internal/logger) to allow for the logging implementation to be replaced. However, this also introduced a circular dependency between the core v23 module and the logger module. This prevents a smooth upgrade to v0.1.10 using go get. This PR breaks that circularity and a new release will be cut that contains it.