x/ref/{cmd,lib}/vdl: changes to work with overlapping module paths (#163)
Some go modules and github repo's are set up in such a way that go.mod module name and the directory structure overlap. For example, github.com/grailbio/base is often cloned into a directory <some-path>/base. In order to write the vdl output to the correct directory under base the vdl compiler needs to allow for the module name github.com/grailbio/base overlapping with base/path-to-package.