v23/context: ensure that values are available across go/v context translations (#167)
This PR ensures that any values stored in a vanadium or go context, that is then manipulated as a go context and translated back into a vanadium context are available in the final context. See context.TestFromGoContextPrincipal for an example. The solution is to implement context.Value on *v23/context.T and have it chain the call to Value to its parent context.