x/ref/runtime/internal: use aws imds v1 first and then v2 (#174)
AWS' instance meta data service cannot be accessed from a docker container running in bridge mode. This PR changes the behaviour to use the v1 service first and only if that fails, to use the v2 service. This fallback is required since it's possible to configure aws instances to allow v2 only.