release v0.1.16 (#188)
Prior to this PR, a server using the 'all' proxy policy would expand the set of proxy instances that it used when it encountered an error with any of existing proxy instances. This PR changes the behavior so that servers will monitor the set of available proxy instances and expand to use all available ones ahead of any errors being encountered.

Also, include a fix for big sur whereby the error messages for connection refused appears to have changed.