v23/vdl,verror: clean up the registration and handling of native types that are interfaces. (#183)
vdl allows for the registration of native types and their conversions from/to wire types. verror uses this facility to convert from vdl.WireError to the go error type, using verror.E as the concrete type. This process is complicated because the native type is an interface and when being decode from an optional value (i.e. pointer), the decoder does not know what concrete type to use. In the past verror was special cased, whereas this PR introduces a general mechanism for specifying the concrete type to use. This simplifies the code somewhat and makes the underlying problem easier to appreciate.