v23/vom: remove the use of a goroutine within TypeDecoder (#184)
vom.TypeDecoder used a goroutine to asynchronously read from the type stream whilst offering up an API for the Decoder to access the types so found and built. This goroutine blocks on the type stream and hence cannot terminate until that stream is closed. This makes it both inefficient and impossible to reliably reuse the TypeDecoder. This PR changes TypeDecoder to synchronously read from the type stream when a new type is found in the value stream for that type. The TypeDecoder is concurrent so care has to be taken to ensure that only goroutine is reading/building types at a time. This new structure ensures that TypeDecoders are both more efficient and can be reused. This allows for the single shot decoding to reuse with expected perf gains in terms of time and memory allocations. The new code also offers a simpler API without the need for Start/Stop methods and complex synchronization.