v23/vdl: additional improvements (#189)
This PR adds a structured cache for additional performance improvements. Multiple performance related caches are grouped into a single structure with an API designed to allow for multiple items to be retrieved simultaneously and be filled incrementally as needed rather in one shot. Benchmarks will be included at the end of this series of PRs, but this PR provides the most benefit for large/complex structs on the reflect path. Caches are maintained for the results of checking that a value's type implements a particular interface. The results of looking up 'nativeInfo' are cached along with the 'implements' info. The 'field' to 'index' lookup is reorganized to return the map for the entire struct rather than a field at a time.