v23/vdl: replace async type conversion with a synchronous version (#190)
This PR replaces the use of the 'pipe' asynchronous converter with a synchronous implementation. This new approach consists of encoding to a vdl.Value and then decoding from that vdl.Value using the existing decoder provided by vdl.Value. The vast bulk of the new code is in toValueEncoder which implements vdl.Encoder. The resulting code is much more efficient for small types (the common case) and avoids the synchronization and storage allocation overheads of the previous approach. The differences are easily seen in profiles. Complete benchmarks will be provided at the end of this series of PRs.