v23/vdl: add support for calling 'iszero' and 'conversion' functions directly (rather than via reflect.Call). (#191)
Prior to this PR the checks for a native value being zero and wire<->native conversions would use reflect.Call to perform the conversions, and for the zero checks would create a wire type from a native type (using reflect.Call) and then test the wiretap for being zero. This PR adds the ability to optimize this by testing to see if the native type implements an interface, currently 'IsZero' and to bypass the reflect.Call mechanism by having a 'converter' type registered that can be called directly rather than using reflect.Call. This are useful optimizations for time.Time and vdl.WireError which are otherwise annoyingly expensive.