v23/vdl: add benchmark results (#192)
This PR includes benchmark results to show the effect of the PRs #182, #183, #184, #189, #190 and #191. See the delta- file for the % improvements. The most dramatic improvements are for the cases that involve native<->wire conversions and reflection based tests. Overall, conversion and type decoding is now synchronous, more memory efficient and generally more streamlined. There are still opportunities for significant improvements for both generated and reflection code, but those are deferred to a later time.