x/ref/lib/slang: a simple, type checked, scripting language (#200)
This PR adds a simple, type checked scripting language. It is essentially a subset of go that supports only function invocation. The initial use case will be for managing security credentials which is in a forthcoming PR. The scripts look as follows:

  p := usePrincipal("$V23_CREDENTIALS")
  printPrincipal(p)

The intent is to avoid multiple, wordy invocations of command line tools to create simple principals and blessings. Although the initial use case is authentication, slang itself is general since the functions that can be invoked are defined by its clients.