x/ref/lib/aws/vxray: xray enabled vtrace implementation (#201)
This PR adds support for using AWS X-ray to save distributed trace data via the vxray package. It required various changes
to the vtrace API and a refactoring of its implementation to allow for common components to be shared across
different implementations.

To understand how to use the support please look at x/ref/examples/echo which has been updated to use the foray implementation of vtrace. A docker compose file is provided to show to run the various components and with appropriate credentials to push xray segments to AWS.

Using xray backed vtrace is transparent other than requiring initialization:

ctx, _ = vxray.InitXRay(ctx, v23.GetRuntimeFlags().VtraceFlags, xray.Config{ServiceVersion: ""}, vxray.EC2Plugin(), vxray.MergeLogging(true))

Once enabled, use of the vtrace API will result in xray traces being generated provided and pushed to the xray daemon as per X-ray's usual configuration. It is possible to mix in native calls to xray since the vxray package strives to use the same context keys as that package to integrate with its use of the context. For example, it is possible to use X-rays HTTP handler as an entry point to create a new trace and then associate all nested RPCs made by that handler with it.