x/ref/lib/aws/vxray: add support for obtaining eks/k8s data (#206)
This PR adds support for obtaining the eke/kubernetes(k8s) cluster name, container id etc. Note, this is PR is not specific to EKS and can be used for other k8s systems. It relies on the convention that the k8s cluster is configured to write a configmap with the cluster's name as one of the key/value pairs. The vxray setup allows for both the name of the configmap and the key for the cluster name to be configured.