v23/vom,v23/rpc,v23: additional debugging, more conservative locking (#234)
The decode tests in v23/vom fail very rarely for unknown reasons on slow CI systems, this PR adds additional debugging to help track that down.

The state transitions for an RPC server are racy in the sense that there is a lot of code between the server being created, initializing itself and being read to accept requests. This seems to show up in some tests where the client code attempts to connect to the server while it is still initializing. This PR changes the states to Initialiazing, Ready, Stopping, Stopped from Active, Stopping and Stopped. A testutil function WaitForServerReady is provided for clients to wait for a server to be ready before attempting to use it, though this only really an issue for tests.

Similarly, the proxyd tests sometimes fail with a panic on slow systems and the change to v23/model.go uses more conservative locking to prevent these panics.