x/ref/lib/security,x/runtime: fix flaky tests, including TestDaemonMode and overly tight timeouts in RPC tests. (#235)
This PR fixes a flaky test by waiting for the expected changes to take place. Issue #236 calls for a better mechanism for doing so.

Tight timeouts are extended in the rpc tests.

Also, add the O_CREATE flag when creating read-only locks to cleanup unnecessary complexity in the code that uses the internal lockedfile package.