chore: update dependencies for macOS Monterey and go 1.17.5 (#240)
Updates for macOS Monterey and packages as of go 1.17.5. The stats/sysstats package is no longer directly imported by the runtime and the version of gopsutils is now at a specific branch/PR that contains a patch for Monterey (which will be updated once that PR is merged and a new release is made). This allows for selective use of sysstats which avoids the unnecessary overhead of sysstats for command line tools, clients etc. The ounttable and the debug servers are the only services to enable sysstats using an import as follows:

import _ "v.io/x/ref/lib/stats/sysstats"