v23/security: extend BlessingRoots API to allow for recognizing x509 derived certificates. (#245)
This PR adds a RecognizedCert method to the BlessingRoots API that is intended to replace the current Recognized method. The latter accepts a public key directly, whereas RecognizedCert accepts a security.Certificate which in the future will include an x509 certificate in addition to the current vanadium contents with the intention of allowing x509 certificates to be verified directly without the need to explicitly add the public keys of all web recognized Certificate Authorities to the blessing roots.