x/ref/test/sectestdata: add more keys and rationalize API (#253)
This PR adds a more complete of keys to sectestdata and also cleans up the API to make it easier to write tests that cover all key types, including ssh and ssl keys. The creation of ssh agent managed private keys is moved to lib/security/ssh to allow it boy accessed by the sectestdata package. A subsequent PR will clean up how the public keys for ssh are handled.