x/ref/runtime/internal/rpc/test: add ED25519 and RSA signatures to the RPC protocol. (#260)
This PR adds the ability to use ED25519 and RSA keys to authenticate RPC. This requires extending the RPC auth message to carry the new signature types. This is done in a backwards compatible manner.

Tests are included for the expected errors for different combinations of crypto keys against the v0.1.20 release. ED25519 keys used by a server will lead to a client side protocol error. RSA keys cannot be loaded by earlier releases and such applications will fail when loading a principal that contains an RSA key as either a private key or when loading a RSA public key added as a blessing root.