x/ref/lib/security/keys/sshkeys: fix racy test (#265)
There was a rare race between zeroing out the passphrase stored in the ssh agent signer implementations via its finalizer and it being used by the signer implementation in a deferred function. This PR simplifies this logic and hopefully avoids the race by not having the signer zero out the passphrase, rather, it keeps a reference to the HostedKey which creates the signer and accesses the passphrase directly from that key.