x/ref/lib/security/keys/sshkeys: fix racy test again (#270)
It seems too complicated to zero the agent passphrase automatically without creating a race. Consequently, the sshkeys package no longer attempts to zero the passphrase. This is not so bad since the passphrase is used to unlock/lock the ssh-agent and can be changed easily and frequently and is, in any case, only of value when agent forwarding is used. As more experience is gained with using different types of agent (eg. ones hosted in a Secure Enclave) the handling of the passphrase may also change.