x/ref/lib/security: allow for LoadPrincipalOpts to use custom blessing store and blessing roots. (#274)
The existing API for loading an existing principal did not allow for using a custom blessing store or blessing root due to the need for those stores to have access to the key information to allow for signing/verification changes to the stores. And since this information is not available prior to loading or outside of the principal it's not possible to create a store externally. This PR fixes this by allowing for a function to be specified as an option to LoadPrincipalOpts which is invoked with the appropriate signer and/or public key to allow for custom stores to be created. It also allows for an x509 certificate to be extracted from a public key that is stored as a certificate.