x/ref/lib/discovery/plugins/ble: stop the discovery Trigger. (#295)
This PR stops the discovery Trigger when the ble plugin is stopped. This bug was noticed while debugging the flaky failures of the ble tests. This fix should reduce the number of outstanding goroutines when the tests do fail. Additional logging is added to help track down the cause of the failures which take the form of the tests running for longer than 10 minutes which suggests some kind of deadlock in the tests or the ble code since the tests generally run in .1 or .2s.