v23/naming,v23/flow/message: reduce mem allocations of using Endpoints (#298)
This PR uses strings.Builder to implement naming.Endpoint.String() to reduce the number of allocations required and to improve its performance. Similarly the implementation of the flow/message is improved to avoid unnecessary copies when serializing and parsing endpoints. Appending a setup message now requires 11 allocations vs 31 before this PR.