v23/flow/message,v23/naming: add message.CopyBuffers and remove naming.ParseEndpointBytes. (#299)
This PR adds message.CopyBuffers in anticipation of experimenting with using shared buffers for network input. CopyBuffers would be called to make copies of any internally use buffers that refer to the original buffer that the message
was read from. ParseEndpointBytes is also removed since it proved too hard to manage the lifecycle of endpoints. Finally, the order in the switch statement in message.Read is modified to have more commonly used messages first.