v23/flow/message: improve the performance of the varint read/write routines. (#301)
This PR improves the performance of the varint read/write routines by unrolling loops. The new routines are about 3x faster and are faster for shorter as opposed to longer ints. To see the differences run the benchmarks: go test -bench=. - the previous code is retained for backwards compatibility testing.