x/ref/runtime/protocols/lib/framer: bound memory usage (#302)
This PR ensures that the framer implementation will never use more than a fixed amount of memory. The previous implementation would always copy the supplied data to a newly allocated buffer in order to prepend the frame size before issuing a single system call to write the data. The new implementation instead will only do this for small payloads, for larger ones it will issue multiple system calls, one for the message size and then one per payload argument. This also reduces the amount of data copying which should be beneficial for large payloads.