x/ref/runtime/internal/flow/conn: minor changes and comments for the readq implementation (#304)
This PR tidies up and documents the implementation of the readq used to pass data from a flow's network receiver code to the flow's Read and ReadMsg methods used by higher level code. The current implementation assumes that buffers may be passed directly from the network to the readq and then on to the caller of GetMsg (called on a different goroutine) and this is now clearly documented/defined.