x/ref/runtime/internal/flow/conn: add the ability to frame outbound messages in messagePipe.writeMsg (#310)
This PR adds the ability to frame messages directly in the messagePipe.writeMsg implementation. Prior to this the framer.WriteMsg is used which either:

issues two system calls, one for the size and a second one for the payload, or
allocates a new buffer leaving space for the frame size and then copies the payload after it
With this PR, messagePipe.writeMsg can leave room for the frame size in the buffers it allocates and then write the size into the start of the buffer when it knows the size of the payload, this buffer can then written directly to the network via the framers Write method.

To make this work, two somewhat ugly implementation choices are required:

1. define framer.T as an explicit interface for framer's to use that includes the full set of methods needed to implement the above strategy.
2. the framer.T must be exposed by the flow.Conn instance created by the TCP protocol implementation.
3. the implementation of messagePipe.writeMsg must test for this interface and then act accordingly.
4. the implementation of messagePipe.writeMsg must test to ensure that the buffer used by the cipher.Seal operation is indeed the one that it specified. This is achieved by testing that the two slices involved share the same underlying storage by comparing the address of the last elements ie. &x[0:cap(x)][cap(x)-1] == &y[0:cap(y)][cap(y)-1].

This results in saving one memory allocation and an associated speedup in using the messagePipe as can seen from the included benchmarks.