v23/vdl: reduce memory allocations and speedups (#316)
Replace the use of small maps for detecting cycles with slices since they are both more space efficient and faster. Also ensure that the buffers used when creating the 'typecons' names for new types can be allocated on the stack which required a small refactoring. These improve the performance of the blessings/discharge exchanges when creating new connections. Benchmarks are included and also show that the use of a union is expensive as compared to structs. Changing the blessings protocol to use structs will be carried out in the near future and the RPC version changed accordingly since it's a backwards incompatible change.