x/ref/runtime/internal/flow/conn: refactor flow control stats (#323)
This PR separates out the flow control counters from the core Conn and Flow structs, moving into their own data types with specific locking for the flow control stats. This is a step towards simplifying the current locking strategy which relies on a single lock on the Conn to guard both the Conn and Flow's core data structures. The only shared state between a Conn and Flow are the flow control counters and by breaking this out it will allow for the simplification of both the Conn and Flow's locking going forward