x/ref/runtime/internal/{flow/conn,rpc/test}: attempt to fix flaky test (#324)
This PR attempts to fix the flaky RPC tests TestTimeoutResponse and TestStreamTimeout which sometimes fail with an EOF rather a timeout error. The attempted fix is in conn.readq.waitLocked which will now return a context.DeadlineExceeded error when the context contains that error rather than converting all context errors to EOF. Additional changes are required to unwrap the timeout error and act accordingly. In addition, the tests themselves are run with different timeouts and additional vtrace output to further diagnose the problem if it persists.