x/ref/runtime/internal/flow/conn: tidy up tracking of unopened to opened states for flows. (#328)
Flows may be created implicitly when accepted, or explicitly when dialed. A dialed flow must send an OpenFlow message to its peer, whereas an accepted flow is created in response to receiving an OpenFlow message. This PR consolidates the handling of these two cases and in particular the communication to the underlying Conn object's waitgroup that an explicitly created, dialed, flow has sent its OpenFlow message.