x/ref/runtime/internal/flow/conn: refactor the 'writeq' and associated locking (#322)
This PR reimplements the 'writeq' used to order and prioritize access to the connections underlying message pipe. The new writeq implementation is significantly simpler to understand, in particular, the locking of the conn and notification via the writeq are no longer intertwined. The dual use of the writeq to prioritize and queue writes as well as to wait for flow control tokens is now separated. Waiting for flow control tokens is now implemented with a separate channel and not via the writeq which allows for significant simplifications to flw.writeMsg. Locking is now more granular and is used purely for ensuring safe access to shared state and not for scheduling of writes.