x/ref/runtime/internal/flow/conn: fix race condition in the writeq implementation (#347)
This PR addresses a race condition in the writeq implementation between normal notification of writeq writers and cancelations of those writers. When cancelations are mixed in with normal operation the cancelations could lead to situations where the writeq invariants are violated. In particular, a writer may have both a normal notification and a cancelation in flight and the receiving select statement may chose one or the other. If it choses the cancelation then its necessary to drain the notification channel. Similarly, notifications are sent whilst holding the writeq's lock to ensure that only one notification may be sent at a time.