x/ref/runtime/internal/flow/conn: refactor how 'remote borrowing' is tracked (#348)
This PR refactors the flow control data structures to avoid the need to use a map to track which flows are using borrowed tokens on their remote ends to reduce the storage/gc overhead of maintaining this map. It moves the state for tracking which flows have remote ends that are borrowing tokens to flowControlFlowStats from flowControlConStats.

It also clarifies the fact that the locking order should be connection first and then flow control.