x/ref/runtime/internal/flow/conn:  refactor how 'toRelease' is tracked and Release messages are constructed (#349)
This PR refactors the flow control data structures to track how many tokens are to be released to a flow's remote end. In addition it refactors how the Release flow control message is created. Both changes reduce the need to use a map and the associated storage/gc overhead. In addition, constructing the release message themselves is simplified.

* .