x/ref/runtime/internal/flow/conn: minor cleanup of 'outstanding borrrowed' flow control (#350)
This PR simplifies and clarifies the logic around the handling of 'outstanding borrowed' tokens, setting the stage for reworking the scheme for handling the return of borrowed/shared flow control tokens when flows are closed. In addition, the allocation of the map used on the client side is differed until it is actually required. Finally, the flow control counters test in close_test.go is moved to flowcontrol_test.go and the invariant tests extended.