x/ref/runtime/internal/flow/conn: add netBuf type (#352)
This PR introduces netBuf's which provide the ability to allocate buffers from either a set of sync.Pools of varying sizes or from the heap. The goal being to reduce the overall memory footprint of sending/receiving messages by allocating buffers that more closely match the size of the messages themselves. The prior behavior was to always allocate buffers of the maximum possible required size. To make use of this the various different types of message are separated out and serialized directly in order to determine their size prior to allocating buffers before transmission. Subsequent PRs will use netBuf's more message reception.