x/ref/runtime/internal/flow/conn: refactor BufferingFlow to use netBufs (#353)
This PR reduces the memory and copying overhead of the BufferingFlow implementation. It reduces the size of the embedded array with a fallback to use a netBuf if additional buffering is required (up to, but not beyond the mtu size). It also fixes a bug whereby payloads larger than the mtu are written out immediately rather than being buffered. The tests are extended to cover all cases and errors.