v23/flow/message,x/ref/runtime/internal/flow/conn: replace the slice of buffers payload with just a buffer (#354)
This PR modifies message.Data and message.OpenFlow to carry a single buffer (rather than a slice of buffers) for their payload in order to reduce the number of allocations and GC overhead. Arguably, this is a breaking API change but in practice the message package is only used by x/ref/runtime/internal/flow/conn. Note that the wire protocol requires that the slice of buffers be written out as a single payload. Therefore this change has no effect beyond reducing memory allocations and simplifying the code. Prior to this change all message's received always used only a single buffer for their payload as dictated by the wire protocol; hence the slice of buffers was purely an overhead for message reception. For transmission of Data and OpenFlow messages the flow implementation is required to reformat variadic buffer arguments to WriteMsg according to the available flow control tokens and hence it is reasonable for it to always reformat to a single buffer if it so chooses.