x/ref/runtime/internal/flow/conn: refactor and rename 'popFront'. (#357)
This PR reimplements and renames 'popFront' as 'removeAtMost' to avoid unnecessary memory allocations. Instead of creating a new slice of byte slices it returns an index into the slice of slices and an offset into the current slice. This leads to many fewer allocations for large (greater than MTU) sized payloads. A 1MB message now needs 298 vs 328 allocations and a 10M message 980 vs 1282.