v23/flow: use a slice rather than map for representing Discharges. (#371)
Discharges are generally small in number and hence a slice is a more appropriate representation for them than a map. The map was originally used to avoid writing a simple 'Find' function but this has led to conversion to/from maps/slices. This PR uses a slice representation throughout but otherwise has no impact on performance or memory allocations.