x/ref/runtime/internal/flow/conn: cleanup up readRemoteAuthLoop (#372)
This PR cleans up the logic and documentation for readRemoteAuthLoop. In particular it makes it clear how blessings are formatted as messages and hence how they should be read in terms of those messages. This sets the stage for refactoring how blessings are encoded in a future version of the RPC/flow protocol since this is currently a fairly inefficient portion of the protocol.